Split profiler into separate package
