+++ utils/profiler/README.md
@@ -1,0 +1,9 @@
+# `@lerna/profiler`
+
+> An internal Lerna tool
+
+## Usage
+
+You probably shouldn't, at least directly.
+
+Install [lerna](https://www.npmjs.com/package/lerna) for access to the `lerna` CLI.
